Rework VideoStream slightly
Clips are now provide a little more access to their assosciated
VideoStreams. This was needed in order to clean up the VideoFrame
features a little bit. As a result, videos now play instantly on
load, so the next step is attaching playback features to the
avVideo.

Add setSource to VideoFrame
Fix frame movement when zoomed in on VideoFrame
Break tracker scrubbing (it has no VideoStream in its ghost clip)